update useDomainFromAddress & useDisplayName
